Move tests for VersionError into own module
